Fixed daemon double update glitch.
If the player was in either item state, the daemon update would also be
called. This change moves the daemon update calls to inside the movement
state. There are also daemon calls added to the knockback part of the code so
things like bullets don't stop moving when the player is getting hurt.